+++ CHANGES.md
@@ -11,7 +11,12 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
-CHANGES  4.4.0 17/04/2025
+CHANGES  4.4.0.? 17/04/2025
+---------------------------
+    * added force option to ttfonts.shapeStr/shapeFragWord
+    * fix wrong lengths in bidiShapedText
+
+CHANGES  4.4.0   17/04/2025
 ---------------------------
     * merge bidi version --> 4.4.0
     * preliminary support for glyph shaping in south Asian languages